refactor: extract rendering and layout hooks
